3/3

+++ src/Q3/Prog703s/Computers.java
@@ -2,16 +2,16 @@ package Q3.Prog703s;
 
 public class Computers {
     String myName;
-    int myValue;
+    double myValue;
     int myNumber;
 
-    public Computers(String name, int number,int value) {
+    public Computers(String name, int number,double value) {
         myName = name;
         myValue = value;
         myNumber = number;
     }
 
     public String getName() { return myName; }
-    public int getValue()   { return myValue; }
+    public double getValue()   { return myValue; }
     public int getNum()     { return myNumber; }
 }
