+++ src/Q3/Prog703s/Linux.java
@@ -3,7 +3,7 @@ package Q3.Prog703s;
 public class Linux extends Computers{
     int hiddenCode;
 
-    public Linux(String name, int number, int value, int code) {
+    public Linux(String name, int number, double value, int code) {
         super(name, number, value);
         hiddenCode = code;
     }
