+++ src/Q3/Prog703s/Mac.java
@@ -3,7 +3,7 @@ package Q3.Prog703s;
 public class Mac extends Computers{
     String myColor;
 
-    public Mac(String name, int number, int value, String color) {
+    public Mac(String name, int number, double value, String color) {
         super(name, number, value);
         myColor = color;
     }
