+++ src/Q3/Prog703s/Windows.java
@@ -3,7 +3,7 @@ package Q3.Prog703s;
 public class Windows extends Computers{
     double myVersion;
 
-    public  Windows(String name, int number, int value, double version) {
+    public  Windows(String name, int number, double value, double version) {
         super(name, number, value);
         myVersion = version;
     }
